Reto5/Servidor/requirements.txt
Flask==2.2.3
PyJWT==2.7.0
Werkzeug==2.2.3
